+++ rust-src/external_functions.rs
@@ -42,7 +42,7 @@ pub fn deserialize_state(
 ) -> String {
     match deserialize_state_aux(contract_name, state_bytes, schema) {
         Ok(s) => s.to_string(),
-        Err(e) => format!("unable to get deserialize State due to: {}", e),
+        Err(e) => format!("{}", e),
     }
 }